src/raspi/requirements.txt
# Raspberry Pi Client Requirements

# Core dependencies
Pillow>=10.0.0
requests>=2.31.0
toml>=0.10.2

# For standalone mode (optional - use backend API instead)
FlightRadarAPI>=1.3.23
geopy>=2.4.1

# Raspberry Pi hardware (only needed on actual hardware)
# Uncomment these on Raspberry Pi:
# RPi.GPIO>=0.7.1
# spidev>=3.6
